refactor: adds user_id to the AsyncStorage, Feed displays reveresed items  #6
